v2.2.1: audit fixes (kernels, parsers, offline API), AQ2 DLL, gpu-smi auto-bundle
Kernels: wave_attn/wave_attn_long missing K->V LDS barrier, causal K-tile
clamp, Dim<=128 guards, thread-local + RAII gemv_tune, 64-bit FP8 GEMM index
arithmetic, call_once mxfp4 LUT, guarded IQ2_S dot2 shuffle.
Untrusted input: GGUF string/array bound, legacy HQ2 manifest bound, alignment
check, explicit (non-assert) shape validation, math.prod element counts.
Offline API: imatrix element-count check, ggml_blck_size_for restype.
Hardening: absolute-path _C load (no sys.path mutation), absolute-only
HIP_QUANT_DLL, gpu-smi fetched from GitHub Releases at build time and bundled.
Self-hosted torch 2.15 wheels indexed via b2_index/torch_index.html.

_C rebuilt multi-arch (gfx942/950/1100-1103/1200/1201) for torch 2.9.1;
hip_quantize.dll + hip_quantize_rocm721.dll rebuilt for the AQ2 guard fix.

+++ src/hip_quant/__init__.py
@@ -164,7 +164,7 @@ class HipQuant:
             self._quantize_tensor_hq2_iters = None
         self._dll.ggml_type_size_for.restype = ctypes.c_size_t
         self._dll.ggml_type_size_for.argtypes = [ctypes.c_int]
-        self._dll.ggml_blck_size_for.restype = ctypes.c_size_t
+        self._dll.ggml_blck_size_for.restype = ctypes.c_int
         self._dll.ggml_blck_size_for.argtypes = [ctypes.c_int]
         self._dll.ggml_row_size_for.restype = ctypes.c_size_t
         self._dll.ggml_row_size_for.argtypes = [ctypes.c_int, ctypes.c_int64]
